Added a settings file, it's not yet completely done but still.
Removed some code in the ClientTickEvent

The whole Stock file is revamped, I'm using Objects now instead HashMaps. Still need a list to make it ordered, will try to fix it later.

It automatically imports settings, and a few more things. It's done on the game startup, not on a world join.