I've added a lot of settings, you can now import the settings without reloading the game.
Added a file to read the bal.

Allows you to change the location of the GUI

Round inventory to 2 numbers after the comma if turned on in settings.
Remove text "Total inventory worth: " with a setting
Added a few settings so you can make cobblestone the same amount worth as stone.