Added command to clear the saved chests
Added code so every chest you open gets saved.
If the chest is already opened and saved it shows it at the top left (you can change location in settings file

Added setting to change location of saved chest GUI

Changed code so the location of the settings doesn't matter.
(Not settings file, but the location of the settings in the settings file)
If you've a emtpy row it stops reading.